+++ src/CG.Business/Clients/IClient.cs
@@ -1,0 +1,13 @@
+﻿using System;
+
+namespace CG.Business.Clients
+{
+    /// <summary>
+    /// This interface represents an object that makes calls to a 
+    /// service, or API, to perform work and gather results.
+    /// </summary>
+    public interface IClient
+    {
+
+    }
+}