+++ src/CG.Business/Clients/Options/ClientOptions.cs
@@ -1,0 +1,12 @@
+﻿using CG.Options;
+
+namespace CG.Business.Clients.Options
+{
+    /// <summary>
+    /// This class represents configuration options for a client
+    /// </summary>
+    public class ClientOptions : OptionsBase
+    {
+        
+    }
+}